deprecate-utils: move all utils to core, remove utils package
