--- packages/utils/.gitignore
@@ -1,8 +1,0 @@
-/.nyc_output
-/.tmp
-/dist
-/node_modules
-/temp
-.idea/*
-.vscode/*
-yarn-error.log
