--- packages/utils/README.md
@@ -1,12 +1,0 @@
-# Rattus ORM - utils
-
-**Utils and helpers for other packages**
-
-### Installation
-Use your favorite package manager. For example, yarn:
-```bash
-yarn add @rattus-orm/utils
-```
-
-### Contributing
-Contributions are welcome! Please read our [Contributing Guide](../../CONTRIBUTING.md) for details on our code of conduct, and the process for submitting pull requests.
