--- packages/utils/package.json
@@ -1,76 +1,0 @@
-{
-  "name": "@rattus-orm/utils",
-  "description": "ORM for your JS/TS apps: utils",
-  "homepage": "https://lyohaplotinka.github.io/rattus-orm/",
-  "author": "Lyoha Plotinka",
-  "license": "MIT",
-  "packageManager": "yarn@4.0.1",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/lyohaplotinka/rattus-orm.git"
-  },
-  "bugs": {
-    "url": "https://github.com/lyohaplotinka/rattus-orm/issues"
-  },
-  "keywords": [
-    "orm",
-    "rattus",
-    "utils",
-    "typescript",
-    "javascript"
-  ],
-  "version": "0.0.8",
-  "main": "./dist/all.js",
-  "browser": "./dist/all.mjs",
-  "module": "./dist/all.mjs",
-  "types": "./dist/all.d.ts",
-  "exports": {
-    "./createBasicProviderTest": {
-      "import": "./dist/createBasicProviderTest.mjs",
-      "require": "./dist/createBasicProviderTest.js",
-      "types": "./dist/createBasicProviderTest.d.ts"
-    },
-    "./isUnknownRecord": {
-      "import": "./dist/isUnknownRecord.mjs",
-      "require": "./dist/isUnknownRecord.js",
-      "types": "./dist/isUnknownRecord.d.ts"
-    },
-    "./pickFromClass": {
-      "import": "./dist/pickFromClass.mjs",
-      "require": "./dist/pickFromClass.js",
-      "types": "./dist/pickFromClass.d.ts"
-    },
-    "./vueComputedUtils": {
-      "import": "./dist/vueComputedUtils.mjs",
-      "require": "./dist/vueComputedUtils.js",
-      "types": "./dist/vueComputedUtils.d.ts"
-    },
-    "./sharedTypes": {
-      "import": "./dist/sharedTypes.mjs",
-      "require": "./dist/sharedTypes.js",
-      "types": "./dist/sharedTypes.d.ts"
-    },
-    "./all": {
-      "import": "./dist/all.mjs",
-      "require": "./dist/all.js",
-      "types": "./dist/all.d.ts"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "run -T tsup",
-    "clean": "rm -rf dist && rm -rf node_modules",
-    "lint": "run -T eslint 'src/**/*.ts' --quiet --fix",
-    "typecheck": "bash -c 'tsc -p tsconfig.json --noEmit'"
-  },
-  "devDependencies": {
-    "vitest": "^1.2.1",
-    "vue": "^3.4.15"
-  },
-  "peerDependencies": {
-    "vitest": "^0.34.6",
-    "vue": "^3.0.0"
-  }
-}
