--- packages/utils/src/index.ts
@@ -1,5 +1,0 @@
-export { createBasicProviderTest } from './createBasicProviderTest'
-export { isUnknownRecord } from './isUnknownRecord'
-export { pickFromClass } from './pickFromClass'
-export * from './types'
-export * from './vueComputedUtils'
