+++ tsup.config.base.ts
@@ -12,7 +12,7 @@ export default function createTsupConfig(entries: Record<string, string>, mixinO
     dts: {
       entry: entries,
     },
-    external: ['vue', 'vuex', '@rattus-orm/utils', '@vue/reactivity'],
+    external: ['vue', 'vuex', '@rattus-orm/utils', '@vue/reactivity', 'vitest'],
   }
 
   const merged = merge(configBase, mixinOptions)
